+++ igdectk/auth/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8; -*-
+#
+# Copyright (c) 2015 INRA UMR1095 GDEC
+
+"""
+auth sub-package init.
+"""