+++ igdectk/xml/__init__.py
@@ -3,7 +3,7 @@
 # Copyright (c) 2015 INRA UMR1095 GDEC
 
 """
-Simplest XML encoder.
+xml sub-package init.
 """
 
 from .encoder import Encoder
@@ -11,8 +11,14 @@ from .decoder import Decoder
 
 
 def dumps(obj):
+    """
+    Helper to dumps a Python object to an XML string.
+    """
     return Encoder().encode(obj)
 
 
 def loads(data):
+    """
+    Helper to loads an XML string into a Python object.
+    """
     return Decoder().decode(data)