+++ CHANGES.rst
@@ -1,3 +1,16 @@
 List of changes
 ===============
 
+Version 1.2
+-----------
+
+| 20151130 : Add RESTful handler helper for django form class
+| 20151127 : Improve new RESTful handler
+| 20151126 : Adding support for XML dumps and loads
+| 20151122 : Reorganisation of the sub-packages and modules
+
+Version 1.1
+-----------
+
+Version 1.0
+-----------